Better dependence generation.
We still put both native and bytecode dependences
in the same Makefile.dep file, but now the dependence list
should be more complete.

For example, if you run "make all" first, ocamldep would look
for .ml and .cmo files and when found, it would generate a
.cmx: .cmx dependence, so if you run "make opt" afterwards,
you do not have missing dependencies problem.