Finally, this version works with includes.
Also, the ugly hack of temporary syntax trees is not needed, yet the
processing of macros is sequential so macro redefinitions etc work
properly.

Other things that might be needed, but I'll consider them only if such
a need arise are:
1. Error reporting is probably heavily screwed up because of bogus loc
   fields.  I'll try to solve each problem as it pops.
2. Making everything work also in interface files,
3. Adding toplevel macros so a single macro application will be able to
   create several definitions (if this is implemented then INCLUDE can
   be such a macro with an ML function body).

Tomorrow I'll start trying to move CPP stuff to use this instead.