Do something reasonable for term printing in Ocaml toploop.
Modern versions of Ocaml require installed printers to use the Format library.
The only quick way to do it was to use print a term to a string and use
Format.print_string.

However, it may be a good idea to abandon Rformat and
to use Format instead or to write an alternative implementation of the Rformat
interface using Format.