Changed the wording of the copy_proof wraning. Instead of printing
> copy_proof: warning: rewrites reduce_mul_meta do not match

it will now print

> Copy_proof: warning: rewrite reduce_mul_meta:
>         was recently changed, potential mismatch with an existing proof.

Since people compiling MetaPRL after downloading it for the first time are
likely to see this warning (we almost amost always have mismatches on CVS),
the warning needs to be less confusing.

I am not completely satisfied with the current wording. Can somebody
come up with a better one?