- Moved Filter_prog from filter/base to filter/filter and removed it from the list of modules being linked into the editor/ml/mp.* binaries.
- Also, simplified some of the signatures (replacing functors with plain modules)

Note: we might want to be eventually able to define things on-line from the toploop
and filter_prog might be useful to have linked in for that purpose. But currently
filter_parse is in filter/filter (and not being linked in anyway), so this change
does not make things any worse than they already were.