Adding a version number to cmiz/cmoz/prla/prlb files.
Before this change we had a magic number that was responsible for making sure that
the file has format we expect it to have and that it has a compatible version
of the format. Now there is a separate magic number responsible for the format
(which is not supposed to change, unless we add/remove formats) and the version number
that is supposed to change whenever the format
(or underlying data structures/data representation) changes.