Cleaning up and refactoring the refiner implementation.
Now refiner should keep enough information around to be able
to figure out which axioms a proof is dependent on.

Still to do: need to be able to actually extract out the dependency
information and a UI to print it (including UI for checking whether
teh rule is fully grounded or not).