Added all the documentation printers. To generate documentation, run the command "omake tex". This places the documentation in doc/ps/theories.
This is a new style of documentation: the documentation command
is in the same OMakefile as the theory itself.  It isn't any
longer placed in doc/latex/theories; this directory is depracated.

You will need the latest omake to build.  I just added the LaTeX
generation code.  Also, it only works on Unix for now.

I haven't done the part about generating documentation for all the
theories at once.  I'll do that tomorrow.