An alternative fix to the "InteractiveSummary" problem, hopefully this time it will work.
Package_info would now always set the InteractiveSummary flag (which is
a correct thing to do). And Filter_parse will now force saving to .cmoz
(not just to "first available"), which is a good thing and will force
the InteractiveSummary flag as well (a bit of a hack, but oh well).