We had a nasty problem in Itt_synt_operator where list{...} in an _axiom_ ended up being interpreted as Ocaml!list{...} instead of Itt_list!list{...}!
To avoid such issues in the future, I've changed the types of all the Ocaml
terms to be Ocaml (instead of Term). Ocaml is declared as a subtype of Dform,
but not a subtype of Term. This, BTW, also made the types of the
Ocaml_base_sos operators a bit more descriptive.