(Alexei and Aleksey) We moved out the Itt_rfun module out of the Itt core, turning it into an extension. Now Itt_dfun is a primitive theory and is no longer being derived from Itt_rfun.
The Itt_rfun module and related theories (such as a module containing
the derivation of the dependent functions from Itt_rfun - this module is now
called Itt_dfun_imp) is currently in theories/itt/extensions/rfun. **Note**
this directory is _not_ yet part of the build, but this should be easy to fix
(will do it next; but wanted to commit the core changes fast to reduce the
possibility of conflicts).