The !!!!! in the mk/config message is now !!!
On this topic, exclamations are urgent.  I don't know if it is true for
other English speakers, but to me exclamation points arrest my attention.
So "That's good!" and "That's good." have very different meanings.
Similarly, "That's bad!" means "You idiot!" -- a negative aggressive
statement -- while "That's bad." is conversational.